normalize removed

+++ mblt_model_zoo/vision/utils/preprocess/center_crop.py
@@ -53,4 +53,4 @@ class CenterCrop(PreOps):
             :,
         ]
 
-        return x
+        return x.astype(np.uint8)
